Basic Greep Find, Move & Return
I have added serveral new things to the greep. First I added a boolean
to check if a greep is on a tomato pile, If they are on a tomato pile
its sets the greep memory to the x and y co ordinates and then the greep
will walk towards that location. Once two greeps are on that location
they will load and walk back to the ship.